requirements.txt
# HTTP & parsing
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
html5lib>=1.1

# Data handling
pandas>=1.5.0

# Financial data sources
edgartools>=2.5.0
yfinance>=0.2.22
kafka-python>=2.0.2

# AWS SDK
boto3>=1.26.0

# Airflow
apache-airflow==2.7.1
apache-airflow-providers-amazon>=8.0.0
